Config file changes for building Japanese into /ja/docs
As we internationalize the rest of circleci.com (outer), the Japanese
site will live under circleci.com/ja/. Docs currently lives at `/docs`
and `/docs/ja`, however it's preferable both aesthetically, SEO-wise,
and for ease of linking between the two sites to have Japanese docs
live at `/ja/docs`.

This commit makes necessary config changes to enable a separate build
of the site for the Japanese version, which can be moved into the
desired subdirectory and uploaded to S3 in the right place for the site
to 1) be served at `/ja/docs` and 2) internal links to be consistent.

The next commit takes care of rewriting site-wide links to respect this
change.